+++ dags/process_unbounded_core_dag.py
@@ -8,7 +8,7 @@ This DAG is scheduled to run once, and it triggers a new run with the trigger_ne
 from stellar_etl_airflow.build_file_sensor_task import build_file_sensor_task
 from stellar_etl_airflow.build_load_task import build_load_task
 from stellar_etl_airflow.default import get_default_dag_args
-from stellar_etl_airflow.build_gcs_to_bq_task import build_gcs_to_bq_task
+from stellar_etl_airflow.build_merge_table_task import build_apply_gcs_changes_to_bq_task
 
 from airflow import DAG
 from airflow.operators.bash_operator import BashOperator
@@ -37,9 +37,9 @@ load_trustlines_task = build_load_task(dag, 'trustlines', 'trustlines_file_senso
 The send tasks receive the location of the file in Google Cloud storage through Airflow's XCOM system.
 Then, the task appends the entries in the file to the corresponding table in BigQuery. 
 '''
-apply_account_changes_task = build_gcs_to_bq_task(dag, 'accounts')
-apply_offer_changes_task = build_gcs_to_bq_task(dag, 'offers')
-apply_trustline_changes_task = build_gcs_to_bq_task(dag, 'trustlines')
+apply_account_changes_task = build_apply_gcs_changes_to_bq_task(dag, 'accounts')
+apply_offer_changes_task = build_apply_gcs_changes_to_bq_task(dag, 'offers')
+apply_trustline_changes_task = build_apply_gcs_changes_to_bq_task(dag, 'trustlines')
 
 
 '''
